Start moving the ETH p2p protocol stuff into the trinity module
