+++ p2p/server.py
@@ -50,7 +50,6 @@ from p2p.p2p_proto import (
 from p2p.peer import (
     BasePeer,
     DEFAULT_PREFERRED_NODES,
-    ETHPeer,
     PeerPool,
 )
 from p2p.service import BaseService
@@ -59,6 +58,7 @@ from p2p.sync import FullNodeSyncer
 if TYPE_CHECKING:
     from trinity.db.chain import AsyncChainDB  # noqa: F401
     from trinity.db.header import BaseAsyncHeaderDB  # noqa: F401
+    from trinity.protocol.eth.peer import ETHPeer  # noqa: F401
 
 
 DIAL_IN_OUT_RATIO = 0.75
@@ -81,7 +81,7 @@ class Server(BaseService):
                  base_db: BaseDB,
                  network_id: int,
                  max_peers: int = DEFAULT_MAX_PEERS,
-                 peer_class: Type[BasePeer] = ETHPeer,
+                 peer_class: Type[BasePeer] = None,
                  bootstrap_nodes: Tuple[Node, ...] = None,
                  preferred_nodes: Sequence[Node] = None,
                  token: CancelToken = None,
@@ -94,6 +94,9 @@ class Server(BaseService):
         self.privkey = privkey
         self.port = port
         self.network_id = network_id
+        if peer_class is None:
+            from trinity.protocol.eth.peer import ETHPeer  # noqa: F811
+            peer_class = ETHPeer
         self.peer_class = peer_class
         self.max_peers = max_peers
         self.bootstrap_nodes = bootstrap_nodes
@@ -302,8 +305,8 @@ def _test() -> None:
 
     from p2p import ecies
     from p2p.constants import ROPSTEN_BOOTNODES
-    from p2p.peer import ETHPeer
 
+    from trinity.protocol.eth.peer import ETHPeer  # noqa: F811
     from trinity.utils.chains import load_nodekey
 
     from tests.p2p.integration_test_helpers import FakeAsyncHeaderDB
