+++ trinity/protocol/eth/proto.py
@@ -1,32 +1,36 @@
 import logging
 from typing import (
-    Any,
-    cast,
     List,
     Tuple,
-    TYPE_CHECKING
+    TYPE_CHECKING,
 )
 
-from rlp import sedes
-
-from eth_typing import (
-    BlockIdentifier,
-)
+from eth_typing import BlockIdentifier
 
 from eth.rlp.headers import BlockHeader
 from eth.rlp.receipts import Receipt
 from eth.rlp.transactions import BaseTransactionFields
 
-from p2p.exceptions import ValidationError
 from p2p.protocol import (
-    BaseBlockHeaders,
-    BaseHeaderRequest,
-    Command,
     Protocol,
-    _DecodedMsgType,
 )
 from p2p.rlp import BlockBody
-from p2p.sedes import HashOrNumber
+
+from .commands import (
+    BlockBodies,
+    BlockHeaders,
+    GetBlockBodies,
+    GetBlockHeaders,
+    GetNodeData,
+    GetReceipts,
+    NewBlock,
+    NewBlockHashes,
+    NodeData,
+    Receipts,
+    Status,
+    Transactions,
+)
+from . import constants
 
 if TYPE_CHECKING:
     from p2p.peer import (  # noqa: F401
@@ -34,122 +38,11 @@ if TYPE_CHECKING:
     )
 
 
-# Max number of items we can ask for in ETH requests. These are the values used in geth and if we
-# ask for more than this the peers will disconnect from us.
-MAX_STATE_FETCH = 384
-MAX_BODIES_FETCH = 128
-MAX_RECEIPTS_FETCH = 256
-MAX_HEADERS_FETCH = 192
-
-
-class Status(Command):
-    _cmd_id = 0
-    structure = [
-        ('protocol_version', sedes.big_endian_int),
-        ('network_id', sedes.big_endian_int),
-        ('td', sedes.big_endian_int),
-        ('best_hash', sedes.binary),
-        ('genesis_hash', sedes.binary),
-    ]
-
-
-class NewBlockHashes(Command):
-    _cmd_id = 1
-    structure = sedes.CountableList(sedes.List([sedes.binary, sedes.big_endian_int]))
-
-
-class Transactions(Command):
-    _cmd_id = 2
-    structure = sedes.CountableList(BaseTransactionFields)
-
-
-class GetBlockHeaders(Command):
-    _cmd_id = 3
-    structure = [
-        ('block_number_or_hash', HashOrNumber()),
-        ('max_headers', sedes.big_endian_int),
-        ('skip', sedes.big_endian_int),
-        ('reverse', sedes.boolean),
-    ]
-
-
-class HeaderRequest(BaseHeaderRequest):
-    max_size = MAX_HEADERS_FETCH
-
-    def __init__(self,
-                 block_number_or_hash: BlockIdentifier,
-                 max_headers: int,
-                 skip: int,
-                 reverse: bool) -> None:
-        self.block_number_or_hash = block_number_or_hash
-        self.max_headers = max_headers
-        self.skip = skip
-        self.reverse = reverse
-
-    def validate_response(self, response: Any) -> None:
-        """
-        Core `Request` API used for validation.
-        """
-        if not isinstance(response, tuple):
-            raise ValidationError("Response to `HeaderRequest` must be a tuple")
-        elif not all(isinstance(item, BlockHeader) for item in response):
-            raise ValidationError("Response must be a tuple of `BlockHeader` objects")
-
-        return self.validate_headers(cast(Tuple[BlockHeader, ...], response))
-
-
-class BlockHeaders(BaseBlockHeaders):
-    _cmd_id = 4
-    structure = sedes.CountableList(BlockHeader)
-
-    def extract_headers(self, msg: _DecodedMsgType) -> Tuple[BlockHeader, ...]:
-        return cast(Tuple[BlockHeader, ...], tuple(msg))
-
-
-class GetBlockBodies(Command):
-    _cmd_id = 5
-    structure = sedes.CountableList(sedes.binary)
-
-
-class BlockBodies(Command):
-    _cmd_id = 6
-    structure = sedes.CountableList(BlockBody)
-
-
-class NewBlock(Command):
-    _cmd_id = 7
-    structure = [
-        ('block', sedes.List([BlockHeader,
-                              sedes.CountableList(BaseTransactionFields),
-                              sedes.CountableList(BlockHeader)])),
-        ('total_difficulty', sedes.big_endian_int)]
-
-
-class GetNodeData(Command):
-    _cmd_id = 13
-    structure = sedes.CountableList(sedes.binary)
-
-
-class NodeData(Command):
-    _cmd_id = 14
-    structure = sedes.CountableList(sedes.binary)
-
-
-class GetReceipts(Command):
-    _cmd_id = 15
-    structure = sedes.CountableList(sedes.binary)
-
-
-class Receipts(Command):
-    _cmd_id = 16
-    structure = sedes.CountableList(sedes.CountableList(Receipt))
-
-
 class ETHProtocol(Protocol):
     name = 'eth'
     version = 63
     _commands = [
-        Status, NewBlockHashes, Transactions, GetBlockHeaders, BlockHeaders, BlockHeaders,
+        Status, NewBlockHashes, Transactions, GetBlockHeaders, BlockHeaders,
         GetBlockBodies, BlockBodies, NewBlock, GetNodeData, NodeData,
         GetReceipts, Receipts]
     cmd_length = 17
@@ -189,10 +82,10 @@ class ETHProtocol(Protocol):
         block_number_or_hash if reverse is False or ending at block_number_or_hash if reverse is
         True.
         """
-        if max_headers > MAX_HEADERS_FETCH:
+        if max_headers > constants.MAX_HEADERS_FETCH:
             raise ValueError(
                 "Cannot ask for more than {} block headers in a single request".format(
-                    MAX_HEADERS_FETCH))
+                    constants.MAX_HEADERS_FETCH))
         cmd = GetBlockHeaders(self.cmd_id_offset)
         data = {
             'block_number_or_hash': block_number_or_hash,
