+++ trinity/protocol/les/commands.py
@@ -1,49 +1,34 @@
-from typing import Any, cast, Dict, Iterator, List, Tuple, TYPE_CHECKING, Union
+from typing import (
+    Any,
+    cast,
+    Dict,
+    Iterator,
+    List,
+    Tuple,
+    Union,
+)
 
-import rlp
-from rlp import sedes
+from eth_typing import Hash32
 
 from eth_utils import (
     encode_hex,
     to_dict,
 )
 
-from eth_typing import (
-    BlockIdentifier,
-    Hash32
-)
+import rlp
+from rlp import sedes
 
 from eth.rlp.headers import BlockHeader
 from eth.rlp.receipts import Receipt
 
-from p2p.exceptions import (
-    ValidationError,
-)
 from p2p.protocol import (
-    BaseBlockHeaders,
-    BaseHeaderRequest,
     Command,
-    Protocol,
     _DecodedMsgType,
 )
-from p2p.rlp import BlockBody
-from p2p.sedes import HashOrNumber
 
-from .constants import LES_ANNOUNCE_SIMPLE
-
-if TYPE_CHECKING:
-    from p2p.peer import (  # noqa: F401
-        ChainInfo
-    )
-
-# Max number of items we can ask for in LES requests. These are the values used in geth and if we
-# ask for more than this the peers will disconnect from us.
-MAX_HEADERS_FETCH = 192
-MAX_BODIES_FETCH = 32
-MAX_RECEIPTS_FETCH = 128
-MAX_CODE_FETCH = 64
-MAX_PROOFS_FETCH = 64
-MAX_HEADER_PROOFS_FETCH = 64
+from trinity.protocol.base_block_headers import BaseBlockHeaders
+from trinity.rlp.block_body import BlockBody
+from trinity.rlp.sedes import HashOrNumber
 
 
 class HeadInfo:
@@ -167,44 +152,6 @@ class GetBlockHeadersQuery(rlp.Serializable):
     ]
 
 
-class HeaderRequest(BaseHeaderRequest):
-    request_id: int
-
-    max_size = MAX_HEADERS_FETCH
-
-    def __init__(self,
-                 block_number_or_hash: BlockIdentifier,
-                 max_headers: int,
-                 skip: int,
-                 reverse: bool,
-                 request_id: int) -> None:
-        self.block_number_or_hash = block_number_or_hash
-        self.max_headers = max_headers
-        self.skip = skip
-        self.reverse = reverse
-        self.request_id = request_id
-
-    def validate_response(self, response: Any) -> None:
-        """
-        Core `Request` API used for validation.
-        """
-        if not isinstance(response, dict):
-            raise ValidationError("Response to `HeaderRequest` must be a dict")
-
-        request_id = response['request_id']
-        if request_id != self.request_id:
-            raise ValidationError(
-                "Response `request_id` does not match.  expected: %s | got: %s".format(
-                    self.request_id,
-                    request_id,
-                )
-            )
-        elif not all(isinstance(item, BlockHeader) for item in response['headers']):
-            raise ValidationError("Response must be a tuple of `BlockHeader` objects")
-
-        return self.validate_headers(cast(Tuple[BlockHeader, ...], response['headers']))
-
-
 class GetBlockHeaders(Command):
     _cmd_id = 2
     structure = [
@@ -322,104 +269,6 @@ class ContractCodes(Command):
     ]
 
 
-class LESProtocol(Protocol):
-    name = 'les'
-    version = 1
-    _commands = [Status, Announce, BlockHeaders, GetBlockHeaders, BlockBodies, Receipts, Proofs,
-                 ContractCodes]
-    cmd_length = 15
-
-    def send_handshake(self, chain_info: 'ChainInfo') -> None:
-        resp = {
-            'protocolVersion': self.version,
-            'networkId': self.peer.network_id,
-            'headTd': chain_info.total_difficulty,
-            'headHash': chain_info.block_hash,
-            'headNum': chain_info.block_number,
-            'genesisHash': chain_info.genesis_hash,
-            'serveHeaders': None,
-            'serveChainSince': 0,
-            'txRelay': None,
-        }
-        cmd = Status(self.cmd_id_offset)
-        self.send(*cmd.encode(resp))
-        self.logger.debug("Sending LES/Status msg: %s", resp)
-
-    def send_get_block_bodies(self, block_hashes: List[bytes], request_id: int) -> None:
-        if len(block_hashes) > MAX_BODIES_FETCH:
-            raise ValueError(
-                "Cannot ask for more than {} blocks in a single request".format(
-                    MAX_BODIES_FETCH))
-        data = {
-            'request_id': request_id,
-            'block_hashes': block_hashes,
-        }
-        header, body = GetBlockBodies(self.cmd_id_offset).encode(data)
-        self.send(header, body)
-
-    def send_get_block_headers(self,
-                               block_number_or_hash: BlockIdentifier,
-                               max_headers: int,
-                               skip: int,
-                               reverse: bool,
-                               request_id: int,
-                               ) -> None:
-        """Send a GetBlockHeaders msg to the remote.
-
-        This requests that the remote send us up to max_headers, starting from
-        block_number_or_hash if reverse is False or ending at block_number_or_hash if reverse is
-        True.
-        """
-        if max_headers > MAX_HEADERS_FETCH:
-            raise ValueError(
-                "Cannot ask for more than {} block headers in a single request".format(
-                    MAX_HEADERS_FETCH))
-        cmd = GetBlockHeaders(self.cmd_id_offset)
-        # Number of block headers to skip between each item (i.e. step in python APIs).
-        skip = 0
-        data = {
-            'request_id': request_id,
-            'query': GetBlockHeadersQuery(block_number_or_hash, max_headers, skip, reverse),
-        }
-        header, body = cmd.encode(data)
-        self.send(header, body)
-
-    def send_block_headers(
-            self, headers: Tuple[BlockHeader, ...], buffer_value: int, request_id: int) -> None:
-        data = {
-            'request_id': request_id,
-            'headers': headers,
-            'buffer_value': buffer_value,
-        }
-        header, body = BlockHeaders(self.cmd_id_offset).encode(data)
-        self.send(header, body)
-
-    def send_get_receipts(self, block_hash: bytes, request_id: int) -> None:
-        data = {
-            'request_id': request_id,
-            'block_hashes': [block_hash],
-        }
-        header, body = GetReceipts(self.cmd_id_offset).encode(data)
-        self.send(header, body)
-
-    def send_get_proof(self, block_hash: bytes, account_key: bytes, key: bytes, from_level: int,
-                       request_id: int) -> None:
-        data = {
-            'request_id': request_id,
-            'proof_requests': [ProofRequest(block_hash, account_key, key, from_level)],
-        }
-        header, body = GetProofs(self.cmd_id_offset).encode(data)
-        self.send(header, body)
-
-    def send_get_contract_code(self, block_hash: bytes, key: bytes, request_id: int) -> None:
-        data = {
-            'request_id': request_id,
-            'code_requests': [ContractCodeRequest(block_hash, key)],
-        }
-        header, body = GetContractCodes(self.cmd_id_offset).encode(data)
-        self.send(header, body)
-
-
 class StatusV2(Status):
     _cmd_id = 0
 
@@ -439,40 +288,3 @@ class ProofsV2(Command):
         ('buffer_value', sedes.big_endian_int),
         ('proof', sedes.CountableList(sedes.raw)),
     ]
-
-
-class LESProtocolV2(LESProtocol):
-    version = 2
-    _commands = [StatusV2, Announce, BlockHeaders, GetBlockHeaders, BlockBodies, Receipts,
-                 ProofsV2, ContractCodes]
-    cmd_length = 21
-
-    def send_handshake(self, chain_info: 'ChainInfo') -> None:
-        resp = {
-            'announceType': LES_ANNOUNCE_SIMPLE,
-            'protocolVersion': self.version,
-            'networkId': self.peer.network_id,
-            'headTd': chain_info.total_difficulty,
-            'headHash': chain_info.block_hash,
-            'headNum': chain_info.block_number,
-            'genesisHash': chain_info.genesis_hash,
-            'serveHeaders': None,
-            'serveChainSince': 0,
-            'txRelay': None,
-        }
-        cmd = StatusV2(self.cmd_id_offset)
-        self.logger.debug("Sending LES/Status msg: %s", resp)
-        self.send(*cmd.encode(resp))
-
-    def send_get_proof(self,
-                       block_hash: bytes,
-                       account_key: bytes,
-                       key: bytes,
-                       from_level: int,
-                       request_id: int) -> None:
-        data = {
-            'request_id': request_id,
-            'proof_requests': [ProofRequest(block_hash, account_key, key, from_level)],
-        }
-        header, body = GetProofsV2(self.cmd_id_offset).encode(data)
-        self.send(header, body)
